refactor(shared-ts): move tray-sync unions and payload wire types to @slicc/shared-ts
Fixes #1343. Part of #1294 (P0-1b) — the tray *sync* data-channel
contract joins the tray *signaling* contract (PR #1342) in
@slicc/shared-ts, eliminating the last hand-mirrored protocol copy on
the TS side.

New shared-ts modules (types + pure guards only, platform-agnostic):
- agent-wire-types.ts: AgentEvent, MessageAttachment(Kind),
  ChatMessage, ToolCall, MessageRole, LickState, LickEvent — the
  webapp-owned payload types embedded in the sync unions
- tray-sync-protocol.ts: LeaderToFollowerMessage /
  FollowerToLeaderMessage, TraySyncHelloMessage,
  TRAY_SYNC_PROTOCOL_VERSION, CHERRY_RUNTIME_TAG, summary/target/FS
  types, cherry event messages + guards, CookieTeleportCookie,
  unhandledProtocolMessage

All moved declarations are byte-identical copies (diff-verified per
type). The webapp modules keep their behavior and re-export the types
from the original layer-local sites, so the ~150 webapp-internal
importers are unchanged:
- core/attachments.ts keeps the formatting helpers
- core/agent-types.ts keeps AgentHandle
- scoops/chat-types.ts keeps Session (persistence, not wire)
- scoops/lick-manager.ts keeps the LickManager runtime
- scoops/tray-sync-protocol.ts keeps TraySyncChannel, the
  CDP/snapshot chunking helpers, and the typed factories (they depend
  on TrayDataChannelLike + the webapp logger)

The golden-fixture corpus now imports the unions from the canonical
@slicc/shared-ts site; the regenerated tray-sync-corpus.json is
byte-identical. iOS mirror-pointer comments and the four affected
guides (ios-app/shared-ts/webapp CLAUDE.md, docs/architecture.md) now
point at the shared-ts file. Fixed a stale Swift doc comment still
referencing packages/webapp/src/ui/types.ts (pre-#1324 path).

Verification: all nine typecheck targets green, lint green (incl.
knip), webapp build green, corpus regen no-op, 8,420 webapp +
1,502 other-package tests passing (the single py-realm-shared failure
is the known local pyodide node_modules lag, pre-existing on main).
