+++ packages/shared-ts/CLAUDE.md
@@ -9,6 +9,8 @@ Platform-agnostic primitives shared across `@slicc/webapp`, `@slicc/node-server`
 - `oauth-extra-domains-storage.ts` — pure-JS read/write helpers over the `slicc_oauth_extra_domains` localStorage key. Lets the extension options page (`secrets.html`) share the per-provider extra-domains store with the webapp's `provider-settings.ts` without dragging in the heavier provider module.
 - `sigv4.ts` — pure SigV4 v4 request signer over Web Crypto (`crypto.subtle`), no AWS SDK. Verified against the canonical AWS test vectors in `tests/sigv4.test.ts`.
 - `tray-signaling.ts` — tray signaling wire contract (leader↔worker control messages, follower HTTP bootstrap request/response shapes, bootstrap retry constants). Single source of truth for `@slicc/webapp` and `@slicc/cloudflare-worker`; the iOS follower mirrors a subset in `packages/ios-app/SliccFollower/Models/TrayTypes.swift`. Worker-internal persisted state (`TrayBootstrapRecord`, `TrayRecord`) stays in `packages/cloudflare-worker/src/shared.ts`.
+- `agent-wire-types.ts` — agent/chat wire payload types (`AgentEvent`, `ChatMessage`, `ToolCall`, `MessageAttachment`, `LickEvent`) embedded in the tray sync protocol. Types only; the producing/consuming behavior stays in `@slicc/webapp`, which re-exports them from their original layer-local modules.
+- `tray-sync-protocol.ts` — tray sync data-channel wire contract (`LeaderToFollowerMessage` / `FollowerToLeaderMessage` unions, summary/target/FS types, `TRAY_SYNC_PROTOCOL_VERSION`, `unhandledProtocolMessage`). Partially mirrored by `packages/ios-app/SliccFollower/Models/SyncProtocol.swift` and enforced by the golden-fixture corpus (`packages/webapp/src/scoops/tray-sync-protocol-corpus.ts`). The `TraySyncChannel` runtime stays in `@slicc/webapp` `scoops/tray-sync-protocol.ts`.
 - `bridge-protocol.ts` — standalone/Electron thin-bridge launch contract: `slicc.bridge.v1.` subprotocol prefix, `bridge`/`bridgeToken` query params, `X-Bridge-Token` header, and the hosted origins (`SLICC_HOSTED_ORIGIN`, `SLICC_STAGING_HUB_ORIGIN`) the node-server and extension allowlists compose from. Swift mirror: `packages/swift-server/Sources/Server/BridgeSecurity.swift`.
 - `sign-and-forward.ts` — S3 / Adobe da.live sign-and-forward orchestration (`executeS3SignAndForward` / `executeDaSignAndForward`). Validates an envelope, resolves credentials via an async `SecretGetter`, signs (S3) or attaches a Bearer token (DA), forwards, and returns a JSON-cloneable reply. Consumed by the webapp mount barrel, the node-server Express handlers (via a `SecretStore` adapter), and the extension service worker.
