+++ packages/shared-ts/src/index.ts
@@ -1,5 +1,6 @@
 // Re-exports added in Task 1.2 (secret-masking) and Task 1.3 (secrets-pipeline).
 
+export * from './agent-wire-types.js';
 export * from './base64.js';
 export * from './bridge-protocol.js';
 export * from './cdp-frame-unmask.js';
@@ -12,3 +13,4 @@ export * from './session-secret-store.js';
 export * from './sign-and-forward.js';
 export * from './sigv4.js';
 export * from './tray-signaling.js';
+export * from './tray-sync-protocol.js';
