+++ packages/webapp/CLAUDE.md
@@ -124,7 +124,7 @@ Deep reference: `docs/kernel/process-model.md`.
 ### Tray Sync (multi-browser leader/follower)
 
 - Path: `packages/webapp/src/scoops/tray-*`, plus page wiring in `packages/webapp/src/ui/page-leader-tray.ts` and `packages/webapp/src/ui/page-follower-tray.ts`.
-- `tray-sync-protocol.ts` is the **canonical wire format**. The iOS follower (`packages/ios-app/SliccFollower/Models/SyncProtocol.swift`) mirrors a **subset** — federated `fs.*` and follower-originated CDP/tab.open are TS-only. iOS DOES respond to leader-initiated `cdp.request` / `tab.open` (and sends back `cdp.response` / `cdp.event` / `tab.opened`). See `docs/architecture.md` "Multi-Browser Sync (Tray) Architecture" for the matrix, and `packages/ios-app/CLAUDE.md` for the 5-step protocol-update checklist.
+- `tray-sync-protocol.ts` re-exports the **canonical wire format** (unions + payload types live in `@slicc/shared-ts` `tray-sync-protocol.ts` / `agent-wire-types.ts`) and holds the channel runtime (`TraySyncChannel`, chunking helpers). The iOS follower (`packages/ios-app/SliccFollower/Models/SyncProtocol.swift`) mirrors a **subset** — federated `fs.*` and follower-originated CDP/tab.open are TS-only. iOS DOES respond to leader-initiated `cdp.request` / `tab.open` (and sends back `cdp.response` / `cdp.event` / `tab.opened`). See `docs/architecture.md` "Multi-Browser Sync (Tray) Architecture" for the matrix, and `packages/ios-app/CLAUDE.md` for the 5-step protocol-update checklist.
 - `tray-leader-sync.ts` (`LeaderSyncManager`) — broadcasts agent events, snapshots, scoops list, sprinkle list/content/updates, federated CDP, federated FS; handles inbound requests from followers (snapshot, sprinkle.fetch, sprinkle.lick, scoops.select, CDP/FS routing). `onForwardedLick` callback validates the forwarded lick type against `FORWARDABLE_TO_LEADER`, scrubs any follower-sent origin fields, and stamps `originFollowerId` (bootstrap ID) + `originLabel` (via `labelForFollower(floatType, runtime)`).
 - `tray-follower-sync.ts` (`FollowerSyncManager`) — TS follower used by both the standalone browser follower (`page-follower-tray.ts`) and the per-page `?cherry=1` extension follower iframes. Implements `AgentHandle`, so installing it as the WC chat controller's agent (`wc-chat-controller.ts:setAgent`, via the `setChatAgent` boot callback) forwards user input to the leader instead of a local orchestrator. `forwardLick(event)` sends the generic `lick` follower→leader message (dropped if the channel is closed).
 - **Lick forwarding**: `LickManager.setForwarder(fn | null)` installs a forwarder hook; `dispatch()` is the private chokepoint all emit sites route through. When a forwarder is set AND the lick type is in `FORWARDABLE_TO_LEADER` (currently just `navigate`), the lick is forwarded instead of handled locally. `LickEvent` gained optional `originFollowerId` / `originLabel`, set only by the leader. Standalone floats require a worker↔page bridge (`set-follower-forwarding` / `forward-lick` / `inject-forwarded-lick` messages in `chrome-extension/src/messages.ts`) because the kernel WORKER owns the lick manager while the PAGE owns the sync managers.
